Membuat form request approval, dan sukses menambahkan data request approval ke tabel request approval, namun belum ada tampilan datatables nya

+++ resources/views/index.blade.php
@@ -2,11 +2,11 @@
 
 @section('container')
   <div class="pagetitle">
-    <h1>Stock</h1>
+    <h1>STOCK</h1>
     <nav>
       <ol class="breadcrumb">
         {{-- <li class="breadcrumb-item"><a href="/dashboard/admin">Admin</a></li> --}}
-        <li class="breadcrumb-item active"><a href="/">Transaksi</a></li>
+        <li class="breadcrumb-item active"><a href="/">Stock</a></li>
       </ol>
     </nav>
   </div><!-- End Page Title -->
@@ -74,17 +74,23 @@
                         </td>
                         <td>
                           {{-- <a href="/transaksi/{{ $transaksi->id }}" class="btn btn-info"><i class="bi bi-eye"></i></a> --}}
-                          <button type="button" class="btn btn-warning" id="btn-edit" data-edit-id="{{ $stock->id }}"
-                            data-bs-toggle="modal" data-bs-target="#editStockModal"><i class="bi bi-pencil"></i></button>
-                          <form action="/stock/{{ $stock->id }}" method="POST" class="d-inline delete-form"
-                            id="deleteForm{{ $stock->id }}">
-                            @method('DELETE')
-                            @csrf
-                            <button type="button" class="btn btn-danger"
-                              onclick="deleteConfirmation('{{ $stock->id }}')">
-                              <i class="bi bi-trash"></i>
-                            </button>
-                          </form>
+                          <button type="button" class="btn btn-warning" id="btn-request-approval"
+                            data-stock-id="{{ $stock->id }}" data-bs-toggle="modal"
+                            data-bs-target="#requestApprovalModal"><i class="bi bi-pencil"></i></button>
+                          @if (auth()->user()->is_admin)
+                            <button type="button" class="btn btn-warning" id="btn-edit"
+                              data-edit-id="{{ $stock->id }}" data-bs-toggle="modal"
+                              data-bs-target="#editStockModal"><i class="bi bi-pencil"></i></button>
+                            <form action="/stock/{{ $stock->id }}" method="POST" class="d-inline delete-form"
+                              id="deleteForm{{ $stock->id }}">
+                              @method('DELETE')
+                              @csrf
+                              <button type="button" class="btn btn-danger"
+                                onclick="deleteConfirmation('{{ $stock->id }}')">
+                                <i class="bi bi-trash"></i>
+                              </button>
+                            </form>
+                          @endif
                         </td>
                       </tr>
                     @endforeach
@@ -203,6 +209,70 @@
               <button type="button" class="btn btn-secondary" data-bs-dismiss="modal">Close</button>
               <button type="button" class="btn btn-primary" id="btn-update-"
                 onclick="editConfirmation()">Update</button>
+            </div>
+          </form>
+        </div>
+      </div>
+    </div>
+
+    <!-- Request Modal -->
+    <div class="modal fade" id="requestApprovalModal" tabindex="-1" aria-labelledby="requestApprovalLabel"
+      aria-hidden="true">
+      <div class="modal-dialog">
+        <div class="modal-content">
+          <div class="modal-header">
+            <h1 class="modal-title fs-5" id="requestApprovalLabel">Edit Stock Barang</h1>
+            <button type="button" class="btn-close" data-bs-dismiss="modal" aria-label="Close"></button>
+          </div>
+          <form action="/requestApproval" method="post" id="requestApprovalForm">
+            @csrf
+            <div class="modal-body">
+              <input type="text" class="form-control" id="request-approval-id" name="stock_id" required hidden>
+              <div class="mb-3">
+                <label for="request-approval-name"
+                  class="form-label @error('request-approval-name') is-invalid @enderror">Nama
+                  Barang</label>
+                <input type="text" class="form-control" id="request-approval-name" name="name" disabled
+                  required>
+                @error('request-approval-name')
+                  <div class="invalid-feedback">
+                    {{ $message }}
+                  </div>
+                @enderror
+              </div>
+              <div class="mb-3">
+                <label for="request-approval-amount" class="form-label @error('amount') is-invalid @enderror">Jumlah
+                  Barang</label>
+                <input type="text" inputmode="numeric" class="form-control" id="request-approval-amount"
+                  name="amount" required>
+                @error('name')
+                  <div class="invalid-feedback">
+                    {{ $message }}
+                  </div>
+                @enderror
+              </div>
+              <div class="mb-3">
+                <h6 class="card-text fw-semibold mt-3">Barang Masuk/Keluar</h6>
+                <div class="d-flex justify-content-evenly align-items-center">
+                  <div class="form-check">
+                    <input class="form-check-input" type="radio" name="is_entry" id="request-approval-entry"
+                      value="1" />
+                    <label class="form-check-label" for="request-approval-entry">
+                      Masuk </label>
+                  </div>
+                  <div class="form-check ms-3">
+                    <input class="form-check-input" type="radio" name="is_entry" id="request-approval-outgoing"
+                      value="0" />
+                    <label class="form-check-label" for="request-approval-outgoing">
+                      Keluar </label>
+                  </div>
+                </div>
+              </div>
+            </div>
+            <div class="modal-footer">
+              <button type="button" class="btn btn-secondary" data-bs-dismiss="modal">Close</button>
+              <button type="button" class="btn btn-primary" id="btn-request-approval-"
+                onclick="requestApprovalConfirmation()">Update</button>
             </div>
           </form>
         </div>
@@ -342,8 +412,38 @@
 
             // Tampilkan modal edit
             $('#editStockModal').modal('show');
+          },
+          error: function(xhr) {
+            alert('Gagal mengambil data stock.');
+          }
+        });
+      });
 
+      $(document).on('click', '#btn-request-approval', function() {
+        var stockId = $(this).data('stock-id');
+        console.log("Request Approval for stock ID: " + stockId);
 
+        // Ambil data stock via AJAX
+        $.ajax({
+          url: '/getRequestData',
+          type: 'GET',
+          data: {
+            id: stockId
+          },
+          success: function(data) {
+            // $('#request-approval-id').val(data.id);
+            // $('#request-approval-id').attr('action', '/stock/' + data.id);
+            $('#requestApprovalForm input[name="stock_id"]').val(data.id);
+            $('#requestApprovalForm input[name="name"]').val(data.name);
+            // $('#requestApprovalForm input[name="amount"]').val(data.amount);
+            $('#btn-request-approval-').attr('id', 'btn-request-approval-' + data.id);
+
+            // Tampilkan modal edit
+            $('#requestApprovalModal').modal('show');
+            console.log("Data received for request approval:", data);
+
+            console.log("ID:", $('#request-approval-id').val());
+            console.log("Name:", $('#request-approval-name').val());
           },
           error: function(xhr) {
             alert('Gagal mengambil data stock.');
@@ -417,7 +517,6 @@
           // $('#dynamic-alert').remove();
           $('.pagetitle').after(alertHtml);
 
-          // Refresh tabel stock (ambil ulang data via AJAX)
           refreshStockTable();
         },
         error: function(xhr) {
@@ -524,6 +623,105 @@
           $('.pagetitle').after(alertHtml);
 
           // Refresh tabel stock (ambil ulang data via AJAX)
+          refreshStockTable();
+        },
+        error: function(xhr) {
+          // Hapus alert error sebelumnya jika ada
+          $('#dynamic-alert').remove();
+
+          // Ambil pesan error dari response
+          var message = 'Terjadi kesalahan.';
+          if (xhr.responseJSON && xhr.responseJSON.message) {
+            message = xhr.responseJSON.message;
+          } else if (xhr.responseJSON && xhr.responseJSON.errors) {
+            message = Object.values(xhr.responseJSON.errors).flat().join('<br>');
+          }
+
+          var alertHtml = `
+            <div class="row justify-content-center" id="dynamic-alert">
+              <div class="alert alert-danger alert-dismissible fade show col-lg-12 justify-content-center" role="alert">
+                ${message}
+                <button type="button" class="btn-close" data-bs-dismiss="alert" aria-label="Close"></button>
+              </div>
+            </div>
+          `;
+          $('.pagetitle').after(alertHtml);
+
+          var errors = xhr.responseJSON && xhr.responseJSON.errors ? xhr.responseJSON.errors : null;
+          if (errors) {
+            if (errors.name) {
+              $('#name').addClass('is-invalid');
+              $('#name').next('.invalid-feedback').text(errors.name[0]);
+            }
+            if (errors.amount) {
+              $('#amount').addClass('is-invalid');
+              $('#amount').next('.invalid-feedback').text(errors.amount[0]);
+            }
+          }
+        }
+      });
+    };
+
+    function requestApprovalConfirmation() {
+      Swal.fire({
+        title: "Yakin ingin melakukan penambahan / pengurangan data stock?",
+        text: "Aksi ini akan diinformasikan ke admin untuk dilakukan konfirmasi!",
+        icon: "warning",
+        showCancelButton: true,
+        confirmButtonColor: "#2980B9",
+        cancelButtonColor: "#d33",
+        confirmButtonText: "Yes, update it!"
+      }).then((result) => {
+        if (result.isConfirmed) {
+          console.log("Requesting Approval Confirmed. Trying to request approval to admin...");
+          // document.getElementById('editStockForm').submit();
+          submitRequestApproval();
+        }
+      });
+    }
+
+    function submitRequestApproval() {
+      var form = $(`#requestApprovalForm`);
+      var url = form.attr('action');
+      var formData = form.serialize();
+
+      $.ajax({
+        url: url,
+        type: 'POST',
+        data: formData,
+        headers: {
+          'X-CSRF-TOKEN': '{{ csrf_token() }}',
+        },
+        success: function(response) {
+          console.log("Request Approval success");
+          $('#requestApprovalModal').modal('hide');
+          $('#requestApprovalForm')[0].reset();
+
+          setTimeout(function() {
+            $('.modal-backdrop').remove();
+            $('body').removeClass('modal-open');
+            $('body').css({
+              'overflow': '',
+              'padding-right': ''
+            });
+            $('#requestApprovalModal').removeClass('show').attr('aria-modal', null).css('display',
+              'none');
+          }, 500);
+
+          $('#dynamic-alert').remove();
+
+          var alertHtml = `
+          <div class="row justify-content-center" id="dynamic-alert">
+            <div class="alert alert-success alert-dismissible fade show col-lg-12 justify-content-center" role="alert">
+              ${response.message}
+              <button type="button" class="btn-close" data-bs-dismiss="alert" aria-label="Close"></button>
+            </div>
+          </div>
+        `;
+
+          // $('#dynamic-alert').remove();
+          $('.pagetitle').after(alertHtml);
+
           refreshStockTable();
         },
         error: function(xhr) {
